--- a/logger.go
+++ b/logger.go
@@ -129,22 +129,26 @@ func (log *logger) log(lvl Level, msg string, fields []Field) {
 		return
 	}
 
+	t := _timeNow().UTC()
 	temp := log.Encoder.Clone()
 	addFields(temp, fields)
 
-	entry := Entry{
-		Level:   lvl,
-		Message: msg,
-		Time:    _timeNow().UTC(),
-		enc:     temp,
-	}
-	for _, hook := range log.Hooks {
-		if err := hook(&entry); err != nil {
-			log.InternalError("hook", err)
+	if len(log.Hooks) > 0 {
+		entry := Entry{
+			Level:   lvl,
+			Message: msg,
+			Time:    t,
+			enc:     temp,
 		}
+		for _, hook := range log.Hooks {
+			if err := hook(&entry); err != nil {
+				log.InternalError("hook", err)
+			}
+		}
+		t, lvl, msg = entry.Time, entry.Level, entry.Message
 	}
 
-	if err := temp.WriteEntry(log.Output, entry.Message, entry.Level, entry.Time); err != nil {
+	if err := temp.WriteEntry(log.Output, msg, lvl, t); err != nil {
 		log.InternalError("encoder", err)
 	}
 	temp.Free()