--- a/meta.go
+++ b/meta.go
@@ -88,13 +88,13 @@ func (m Meta) InternalError(cause string, err error) {
 	m.ErrorOutput.Sync()
 }
 
-// RunHooks runs any Hook functions, returning a possibly modified
+// Encode runs any Hook functions, returning a possibly modified
 // time, message, and level.
-func (m Meta) RunHooks(t time.Time, lvl Level, msg string, fields []Field) (Level, string, Encoder) {
+func (m Meta) Encode(t time.Time, lvl Level, msg string, fields []Field) (string, Encoder) {
 	enc := m.Encoder.Clone()
 	addFields(enc, fields)
 	if len(m.Hooks) == 0 {
-		return lvl, msg, enc
+		return msg, enc
 	}
 	entry := Entry{
 		Level:   lvl,
@@ -107,5 +107,5 @@ func (m Meta) RunHooks(t time.Time, lvl Level, msg string, fields []Field) (Leve
 			m.InternalError("hook", err)
 		}
 	}
-	return entry.Level, entry.Message, entry.enc
+	return entry.Message, entry.enc
 }