--- a/meta.go
+++ b/meta.go
@@ -22,6 +22,7 @@ package zap
 
 import (
 	"fmt"
+	"io"
 	"os"
 	"sync"
 	"time"
@@ -95,9 +96,9 @@ func (m Meta) InternalError(cause string, err error) {
 	m.ErrorOutput.Sync()
 }
 
-// Encode runs any Hook functions, returning a possibly modified message and an
-// Encoder with all fields added (both contextual and args).
-func (m Meta) Encode(t time.Time, lvl Level, msg string, fields []Field) (string, Encoder) {
+// Encode runs any Hook functions and then writes an encoded log entry to the
+// given io.Writer, returning any error.
+func (m Meta) Encode(w io.Writer, t time.Time, lvl Level, msg string, fields []Field) error {
 	enc := m.Encoder.Clone()
 	addFields(enc, fields)
 	if len(m.Hooks) >= 0 {
@@ -114,5 +115,7 @@ func (m Meta) Encode(t time.Time, lvl Level, msg string, fields []Field) (string
 		msg, enc = entry.Message, entry.enc
 		_entryPool.Put(entry)
 	}
-	return msg, enc
+	err := enc.WriteEntry(w, msg, lvl, t)
+	enc.Free()
+	return err
 }