This is just the `Entry` prep work out of the tabled #194.