@akshayjshah 2016-12-06
Last nit above. Can you also post the output of the full benchmark suite on master, on this PR, and the `benchcmp` of the two? As we make these refactorings, I'd like to ensure that we haven't mistakenly introduced any perf regressions.