@jcorbin 2016-12-06
No new allocs, and the cpu deltas are within noise (i.e. indistinguishable from two consecutive runs on master; raising `-benchtime` to 10s does not substantially help the noise):
```
$ git checkout master && make bench | tee bench_$(git rev-parse HEAD)
...
$ git checkout imp_entry && make bench | tee bench_$(git rev-parse HEAD)
...
$ benchcmp bench_26c75604fce9f820e6ef1658e0f4bc4f4a6b11a4 bench_945071a5de939921581d80a8686f2e07d75f085c
benchmark                                          old ns/op     new ns/op     delta
BenchmarkApexLogAddingFields-8                     15753         14063         -10.73%
BenchmarkApexLogWithAccumulatedContext-8           12068         11580         -4.04%
BenchmarkApexLogWithoutFields-8                    3061          2623          -14.31%
BenchmarkGoKitAddingFields-8                       4832          3728          -22.85%
BenchmarkGoKitWithAccumulatedContext-8             3880          3580          -7.73%
BenchmarkGoKitWithoutFields-8                      661           635           -3.93%
BenchmarkLog15AddingFields-8                       20491         18031         -12.01%
BenchmarkLog15WithAccumulatedContext-8             17770         17430         -1.91%
BenchmarkLog15WithoutFields-8                      4708          5440          +15.55%
BenchmarkLogrusAddingFields-8                      6776          5090          -24.88%
BenchmarkZapBarkifyAddingFields-8                  2565          2693          +4.99%
BenchmarkLogrusWithAccumulatedContext-8            5034          4910          -2.46%
BenchmarkZapBarkifyWithAccumulatedContext-8        321           260           -19.00%
BenchmarkLogrusWithoutFields-8                     1670          1541          -7.72%
BenchmarkStandardLibraryWithoutFields-8            506           451           -10.87%
BenchmarkZapStandardizeWithoutFields-8             282           267           -5.32%
BenchmarkZapDisabledLevelsWithoutFields-8          3.05          2.70          -11.48%
BenchmarkZapDisabledLevelsAccumulatedContext-8     2.87          2.77          -3.48%
BenchmarkZapDisabledLevelsAddingFields-8           236           226           -4.24%
BenchmarkZapDisabledLevelsCheckAddingFields-8      4.73          4.57          -3.38%
BenchmarkZapAddingFields-8                         886           1029          +16.14%
BenchmarkZapWithAccumulatedContext-8               226           224           -0.88%
BenchmarkZapWithoutFields-8                        220           212           -3.64%
BenchmarkZapSampleWithoutFields-8                  333           288           -13.51%
BenchmarkZapSampleAddingFields-8                   690           631           -8.55%
BenchmarkZapSampleCheckWithoutFields-8             288           326           +13.19%
BenchmarkZapSampleCheckAddingFields-8              304           312           +2.63%
BenchmarkCheckedMessage_Chain-8                    93.0          85.2          -8.39%
BenchmarkCheckedMessage_Chain_sliceLoggers-8       124           82.6          -33.39%
BenchmarkJSONLogMarshalerFunc-8                    139           126           -9.35%
BenchmarkZapJSON-8                                 284           247           -13.03%
BenchmarkStandardJSON-8                            2679          2407          -10.15%
BenchmarkNullLogMarshalerFunc-8                    40.1          42.7          +6.48%
BenchmarkZapNull-8                                 12.5          16.0          +28.00%
BenchmarkNoContext-8                               208           250           +20.19%
BenchmarkBoolField-8                               242           281           +16.12%
BenchmarkFloat64Field-8                            297           288           -3.03%
BenchmarkIntField-8                                250           278           +11.20%
BenchmarkInt64Field-8                              254           298           +17.32%
BenchmarkStringField-8                             236           278           +17.80%
BenchmarkStringerField-8                           268           299           +11.57%
BenchmarkTimeField-8                               247           348           +40.89%
BenchmarkDurationField-8                           263           360           +36.88%
BenchmarkErrorField-8                              253           323           +27.67%
BenchmarkStackField-8                              17494         13891         -20.60%
BenchmarkMarshalerField-8                          423           419           -0.95%
BenchmarkObjectField-8                             1084          882           -18.63%
BenchmarkAddCallerHook-8                           525           482           -8.19%
Benchmark10Fields-8                                648           581           -10.34%
Benchmark100Fields-8                               6879          8108          +17.87%
BenchmarkTee_Check-8                               441           509           +15.42%

benchmark                                          old allocs     new allocs     delta
BenchmarkApexLogAddingFields-8                     64             64             +0.00%
BenchmarkApexLogWithAccumulatedContext-8           49             49             +0.00%
BenchmarkApexLogWithoutFields-8                    11             11             +0.00%
BenchmarkGoKitAddingFields-8                       71             71             +0.00%
BenchmarkGoKitWithAccumulatedContext-8             49             49             +0.00%
BenchmarkGoKitWithoutFields-8                      14             14             +0.00%
BenchmarkLog15AddingFields-8                       91             91             +0.00%
BenchmarkLog15WithAccumulatedContext-8             70             70             +0.00%
BenchmarkLog15WithoutFields-8                      24             24             +0.00%
BenchmarkLogrusAddingFields-8                      79             79             +0.00%
BenchmarkZapBarkifyAddingFields-8                  23             23             +0.00%
BenchmarkLogrusWithAccumulatedContext-8            62             62             +0.00%
BenchmarkZapBarkifyWithAccumulatedContext-8        3              3              +0.00%
BenchmarkLogrusWithoutFields-8                     27             27             +0.00%
BenchmarkStandardLibraryWithoutFields-8            2              2              +0.00%
BenchmarkZapStandardizeWithoutFields-8             3              3              +0.00%
BenchmarkZapDisabledLevelsWithoutFields-8          0              0              +0.00%
BenchmarkZapDisabledLevelsAccumulatedContext-8     0              0              +0.00%
BenchmarkZapDisabledLevelsAddingFields-8           2              2              +0.00%
BenchmarkZapDisabledLevelsCheckAddingFields-8      0              0              +0.00%
BenchmarkZapAddingFields-8                         2              2              +0.00%
BenchmarkZapWithAccumulatedContext-8               0              0              +0.00%
BenchmarkZapWithoutFields-8                        0              0              +0.00%
BenchmarkZapSampleWithoutFields-8                  1              1              +0.00%
BenchmarkZapSampleAddingFields-8                   3              3              +0.00%
BenchmarkZapSampleCheckWithoutFields-8             0              0              +0.00%
BenchmarkZapSampleCheckAddingFields-8              1              1              +0.00%
BenchmarkCheckedMessage_Chain-8                    0              0              +0.00%
BenchmarkCheckedMessage_Chain_sliceLoggers-8       0              0              +0.00%
BenchmarkJSONLogMarshalerFunc-8                    1              1              +0.00%
BenchmarkZapJSON-8                                 0              0              +0.00%
BenchmarkStandardJSON-8                            38             38             +0.00%
BenchmarkNullLogMarshalerFunc-8                    1              1              +0.00%
BenchmarkZapNull-8                                 0              0              +0.00%
BenchmarkNoContext-8                               0              0              +0.00%
BenchmarkBoolField-8                               1              1              +0.00%
BenchmarkFloat64Field-8                            1              1              +0.00%
BenchmarkIntField-8                                1              1              +0.00%
BenchmarkInt64Field-8                              1              1              +0.00%
BenchmarkStringField-8                             1              1              +0.00%
BenchmarkStringerField-8                           2              2              +0.00%
BenchmarkTimeField-8                               1              1              +0.00%
BenchmarkDurationField-8                           1              1              +0.00%
BenchmarkErrorField-8                              1              1              +0.00%
BenchmarkStackField-8                              2              2              +0.00%
BenchmarkMarshalerField-8                          1              1              +0.00%
BenchmarkObjectField-8                             7              7              +0.00%
BenchmarkAddCallerHook-8                           1              1              +0.00%
Benchmark10Fields-8                                1              1              +0.00%
Benchmark100Fields-8                               3              3              +0.00%
BenchmarkTee_Check-8                               1              1              +0.00%

benchmark                                          old bytes     new bytes     delta
BenchmarkApexLogAddingFields-8                     4026          4026          +0.00%
BenchmarkApexLogWithAccumulatedContext-8           2802          2802          +0.00%
BenchmarkApexLogWithoutFields-8                    584           584           +0.00%
BenchmarkGoKitAddingFields-8                       3655          3655          +0.00%
BenchmarkGoKitWithAccumulatedContext-8             2937          2937          +0.00%
BenchmarkGoKitWithoutFields-8                      672           672           +0.00%
BenchmarkLog15AddingFields-8                       5345          5345          +0.00%
BenchmarkLog15WithAccumulatedContext-8             4680          4680          +0.00%
BenchmarkLog15WithoutFields-8                      1479          1479          +0.00%
BenchmarkLogrusAddingFields-8                      5854          5854          +0.00%
BenchmarkZapBarkifyAddingFields-8                  3623          3624          +0.03%
BenchmarkLogrusWithAccumulatedContext-8            4016          4016          +0.00%
BenchmarkZapBarkifyWithAccumulatedContext-8        48            48            +0.00%
BenchmarkLogrusWithoutFields-8                     1465          1465          +0.00%
BenchmarkStandardLibraryWithoutFields-8            32            32            +0.00%
BenchmarkZapStandardizeWithoutFields-8             40            40            +0.00%
BenchmarkZapDisabledLevelsWithoutFields-8          0             0             +0.00%
BenchmarkZapDisabledLevelsAccumulatedContext-8     0             0             +0.00%
BenchmarkZapDisabledLevelsAddingFields-8           704           704           +0.00%
BenchmarkZapDisabledLevelsCheckAddingFields-8      0             0             +0.00%
BenchmarkZapAddingFields-8                         706           706           +0.00%
BenchmarkZapWithAccumulatedContext-8               0             0             +0.00%
BenchmarkZapWithoutFields-8                        0             0             +0.00%
BenchmarkZapSampleWithoutFields-8                  64            64            +0.00%
BenchmarkZapSampleAddingFields-8                   768           768           +0.00%
BenchmarkZapSampleCheckWithoutFields-8             63            63            +0.00%
BenchmarkZapSampleCheckAddingFields-8              66            66            +0.00%
BenchmarkCheckedMessage_Chain-8                    42            42            +0.00%
BenchmarkCheckedMessage_Chain_sliceLoggers-8       42            42            +0.00%
BenchmarkJSONLogMarshalerFunc-8                    16            16            +0.00%
BenchmarkZapJSON-8                                 0             0             +0.00%
BenchmarkStandardJSON-8                            1920          1920          +0.00%
BenchmarkNullLogMarshalerFunc-8                    16            16            +0.00%
BenchmarkZapNull-8                                 0             0             +0.00%
BenchmarkNoContext-8                               0             0             +0.00%
BenchmarkBoolField-8                               64            64            +0.00%
BenchmarkFloat64Field-8                            64            64            +0.00%
BenchmarkIntField-8                                64            64            +0.00%
BenchmarkInt64Field-8                              64            64            +0.00%
BenchmarkStringField-8                             64            64            +0.00%
BenchmarkStringerField-8                           68            68            +0.00%
BenchmarkTimeField-8                               64            64            +0.00%
BenchmarkDurationField-8                           64            64            +0.00%
BenchmarkErrorField-8                              64            64            +0.00%
BenchmarkStackField-8                              1090          1090          +0.00%
BenchmarkMarshalerField-8                          64            64            +0.00%
BenchmarkObjectField-8                             585           585           +0.00%
BenchmarkAddCallerHook-8                           48            48            +0.00%
Benchmark10Fields-8                                642           642           +0.00%
Benchmark100Fields-8                               1169          1169          +0.00%
BenchmarkTee_Check-8                               64            64            +0.00%
```